+++ demo/server/webpackConfig.js
@@ -38,7 +38,13 @@ const makeWebpackConfig = ({
         },
         cache: true,
         devtool: "source-map",
-
+        optimization: {
+            splitChunks: DEV
+                ? {}
+                : {
+                      chunks: "all"
+                  }
+        },
         stats: {
             colors: true,
             reasons: true
@@ -80,7 +86,7 @@ const makeWebpackConfig = ({
                     ]
                 },
                 {
-                    test: /\.(svg|eot|woff2?|ttf|png)$/,
+                    test: /\.(svg|eot|woff2?|ttf|png|pdf)$/,
                     loaders: [
                         {
                             loader: "file-loader"