+++ demo/source/demos/redirects/Pages.js
@@ -33,7 +33,14 @@ class Pages extends Component {
             case "admin":
                 return <AdminPage authenticated={authenticated} />;
             case "content":
-                return <ContentPage slug={slug} content={contentPage} />;
+                // TODO: This worked around a bug (regression test in 04Redirects.js)
+                // navigating from Found Content to Missing Content when it interims
+                // tries to render the Found Content page. Not sure if there's a more
+                // solid approach. Maybe finding a way to keep inbound and outbound
+                // state completely separate.
+                return contentPage ? (
+                    <ContentPage slug={slug} content={contentPage} />
+                ) : null;
             // Note: there is no case for the "moved" page because it never exists
         }
     }