+++ demo/source/demos/redirects/code.js
@@ -1,4 +1,5 @@
 import routes from "!!raw-loader!./routes";
+import api from "!!raw-loader!./api";
 import Root from "!!raw-loader!./Root";
 import Pages from "!!raw-loader!./Pages";
 import Landing from "!!raw-loader!./pages/Landing";
@@ -7,6 +8,10 @@ export default [
     {
         name: "routes.js",
         code: routes
+    },
+    {
+        name: "api.js",
+        code: api
     },
     {
         name: "Root.js",