+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # JARL: Version History
 
+## Next version
+
+*   Fix: Got `active` working everywhere properly, in particular inside demos
+*   Fix: Edge case crash on redirect demos
+*   Demo: Code splitting now loads something substantial and is actually code splitting properly
+*   Demo: Code splitting shows a PDF which can also be navigated via routing
+*   Build: Made Webpack more aggressive about chunks
+
 ## v1.0.0-beta.2
 
 *   New feature: improved API to accessing `resolve` objects. They are now available